Add support for multi-turn MoE router replay by passing an engine-level prompt offset. - Add `routed_experts_prompt_start` to `SamplingParams` and support `Sequence[int]` per-prompt in `VllmSampler` and `InprocessVllmSamplerAdapter`. - Add `env_routed_experts` and `assistant_routed_experts` to `Step` in `agent_types.py` for naming symmetry. - Handle vLLM's autoregressive P + G - 1 routing invariant in `TrajectoryCollectEngine`, dynamically stitching delayed assistant routing during subsequent turn prefills and padding terminal unrouted tokens with `UNSET_ROUTED_EXPERT`. - Expose `env_routed_experts` and `assistant_routed_experts` in `TrajectoryCollectEngine.collect(mode="Steps")` while keeping `routed_experts` globally in `mode="Token"`, alongside strict fail-fast validation. - Keep `routed_experts` as `np.int16` across the Orchestrator pipeline to halve Orchestrator \u2192 Trainer IPC payloads, deferring the `int32` upcast directly to the Trainer loop (`compute_per_token_logps`). - Add comprehensive unit and integration tests (`router_replay_end_to_end_test`, `trajectory_collect_engine_test`, `inprocess_vllm_sampler_adapter_test`, `algorithm_adapter_test`, etc.).
PiperOrigin-RevId: 983364802

+++ tunix/experimental/orchestrator/algorithm_adapter.py
@@ -83,7 +83,7 @@ def _routed_experts_for(
     routed = item.metadata.get("routed_experts")
   if routed is None:
     return None
-  routed_arr = np.asarray(routed, dtype=np.int32)
+  routed_arr = np.asarray(routed, dtype=np.int16)
   if routed_arr.ndim != 3:
     raise ValueError(
         "routed_experts must be [length, num_layers, top_k]; got shape"
@@ -94,7 +94,7 @@ def _routed_experts_for(
   pad = np.full(
       (seq_len - routed_arr.shape[0],) + routed_arr.shape[1:],
       datatypes.UNSET_ROUTED_EXPERT,
-      dtype=np.int32,
+      dtype=np.int16,
   )
   return np.concatenate([routed_arr, pad], axis=0)
